test(keyboard): expose test helpers
