+++ packages/ubuntu_desktop_installer/test/keyboard/keyboard_dialogs_test.dart
@@ -7,7 +7,7 @@ import 'package:ubuntu_desktop_installer/services.dart';
 import 'package:yaru_test/yaru_test.dart';
 
 import '../test_utils.dart';
-import 'keyboard_model_test.mocks.dart';
+import 'test_keyboard.dart';
 
 void main() {
   setUpAll(() => UbuntuTester.context = DetectKeyboardView);
